Move MetadataPanel to UserInterface folder
